requirements/requirements_django.txt
Django==1.9

# Django-extensions
Werkzeug==0.11.3
psycopg2==2.6.1
wheel==0.24.0
django-extensions==1.6.1
django-debug-toolbar==1.4

Pillow==3.1.1

# Jinja2
Jinja2==2.8
django-jinja==2.1.1

django-braces==1.8.1

django-redis==4.3.0

# Restframework
djangorestframework==3.3.2
djangorestframework-camel-case==0.2.0

celery==3.1.20

# Elastic search
elasticsearch==2.2.0
django-haystack==2.4.1

# Tests
django-nose==1.4.3
factory-boy==2.6.1
django-webtest==1.7.8

ramona==1.0.0
gunicorn==19.4.5
